refactor: move src/utils/string.ts to src/lib/stringUtils.ts
Refs CHAOS-1234
