+++ src/lib/stringUtils.ts
